+++ scripts/settings.js
@@ -1,0 +1,16 @@
+import { MODULE_ID } from "./lib/const.js";
+
+export function setupSettings() {
+  game.settings.register(
+    MODULE_ID,
+    "token-magic-fx.ask-disable-template-handling",
+    {
+      name: "",
+      hint: "",
+      scope: "world",
+      config: true,
+      default: true,
+      type: Boolean,
+    },
+  );
+}